requirements.txt
pyttsx3
fastapi
uvicorn
openai
python-dotenv
SpeechRecognition
pyaudio
